trying to get the test harness to work
