+++ CSETWebApi/CSETWeb_Api/CSETWeb_Api.sln
@@ -9,8 +9,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BusinessLogic", "BusinessLo
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CryptoBuffer", "CryptoBuffer\CryptoBuffer.csproj", "{4E44142D-08E2-4F40-9532-896E1F00018C}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CSETWeb_ApiTests", "CSETWeb_ApiTests\CSETWeb_ApiTests.csproj", "{18E1CAF3-160E-49A7-B0FA-52DBBECFED5B}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExportCSV", "ExportCSV\ExportCSV.csproj", "{6F0DD06E-EB8C-4B13-BD4B-20A267F96537}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DataLayerCore", "DataLayerCore\DataLayerCore.csproj", "{EC76CD3A-4B69-4F5C-94A5-32F2E149458F}"
@@ -47,12 +45,6 @@ Global
 		{4E44142D-08E2-4F40-9532-896E1F00018C}.localhost|Any CPU.Build.0 = localhost|Any CPU
 		{4E44142D-08E2-4F40-9532-896E1F00018C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4E44142D-08E2-4F40-9532-896E1F00018C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{18E1CAF3-160E-49A7-B0FA-52DBBECFED5B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{18E1CAF3-160E-49A7-B0FA-52DBBECFED5B}.localdb|Any CPU.ActiveCfg = localdb|Any CPU
-		{18E1CAF3-160E-49A7-B0FA-52DBBECFED5B}.localhost|Any CPU.ActiveCfg = localhost|Any CPU
-		{18E1CAF3-160E-49A7-B0FA-52DBBECFED5B}.localhost|Any CPU.Build.0 = localhost|Any CPU
-		{18E1CAF3-160E-49A7-B0FA-52DBBECFED5B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{18E1CAF3-160E-49A7-B0FA-52DBBECFED5B}.Release|Any CPU.Build.0 = Release|Any CPU
 		{6F0DD06E-EB8C-4B13-BD4B-20A267F96537}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{6F0DD06E-EB8C-4B13-BD4B-20A267F96537}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6F0DD06E-EB8C-4B13-BD4B-20A267F96537}.localdb|Any CPU.ActiveCfg = localdb|Any CPU
