+++ CSETWebApi/CSETWeb_Api/CSETWeb_Api/Web.config
@@ -10,7 +10,7 @@
     
   </configSections>
   <system.webServer>
-    <validation validateIntegratedModeConfiguration="false"/>
+    <validation validateIntegratedModeConfiguration="false" />
   </system.webServer>
   <appSettings>
     <add key="webpages:Version" value="3.0.0.0" />
@@ -217,13 +217,29 @@
         <assemblyIdentity name="System.Web.Mvc" publicKeyToken="31bf3856ad364e35" />
         <bindingRedirect oldVersion="0.0.0.0-5.2.7.0" newVersion="5.2.7.0" />
       </dependentAssembly>
+      <dependentAssembly>
+        <assemblyIdentity name="System.Memory" publicKeyToken="cc7b13ffcd2ddd51" culture="neutral" />
+        <bindingRedirect oldVersion="0.0.0.0-4.0.1.1" newVersion="4.0.1.1" />
+      </dependentAssembly>
+      <dependentAssembly>
+        <assemblyIdentity name="Microsoft.EntityFrameworkCore" publicKeyToken="adb9793829ddae60" culture="neutral" />
+        <bindingRedirect oldVersion="0.0.0.0-2.2.6.0" newVersion="2.2.6.0" />
+      </dependentAssembly>
+      <dependentAssembly>
+        <assemblyIdentity name="Microsoft.EntityFrameworkCore.Abstractions" publicKeyToken="adb9793829ddae60" culture="neutral" />
+        <bindingRedirect oldVersion="0.0.0.0-2.2.6.0" newVersion="2.2.6.0" />
+      </dependentAssembly>
+      <dependentAssembly>
+        <assemblyIdentity name="Microsoft.EntityFrameworkCore.Relational" publicKeyToken="adb9793829ddae60" culture="neutral" />
+        <bindingRedirect oldVersion="0.0.0.0-2.2.6.0" newVersion="2.2.6.0" />
+      </dependentAssembly>
     </assemblyBinding>
   </runtime>
   <connectionStrings>
     <add name="CSET_DB" connectionString="data source=localhost;initial catalog=CSETWeb;persist security info=True;Integrated Security=SSPI;MultipleActiveResultSets=True" />
     <add name="CSETWebEntities" connectionString="data source=localhost;initial catalog=CSETWeb;persist security info=True;Integrated Security=SSPI;MultipleActiveResultSets=True;App=EntityFramework" providerName="System.Data.EntityClient" />
-    <add name="ElmahConn" connectionString="data source=localhost;initial catalog=CSETWeb;persist security info=True;Trusted_Connection=True;MultipleActiveResultSets=True;App=Elmah" providerName="System.Data.EntityClient" />
-    <add name="HangfireConn" connectionString="data source=localhost;initial catalog=CSETWeb;persist security info=True;Trusted_Connection=True;MultipleActiveResultSets=True;App=HangFire" providerName="System.Data.EntityClient" />
+    <add name="ElmahConn" connectionString="data source=localhost;initial catalog=CSETWeb;persist security info=True;Integrated Security=SSPI;MultipleActiveResultSets=True;App=Elmah" providerName="System.Data.EntityClient" />
+    <add name="HangfireConn" connectionString="data source=localhost;initial catalog=CSETWeb;persist security info=True;Integrated Security=SSPI;MultipleActiveResultSets=True;App=Hangfire" providerName="System.Data.EntityClient" />
   </connectionStrings>
   <entityFramework>
     <defaultConnectionFactory type="System.Data.Entity.Infrastructure.LocalDbConnectionFactory, EntityFramework">
