+++ CSETWebApi/CSETWeb_Api/CSETWeb_Api/packages.config
@@ -4,11 +4,11 @@
   <package id="bootstrap" version="4.3.1" targetFramework="net472" />
   <package id="elmah.corelibrary" version="1.2.2" targetFramework="net452" />
   <package id="Elmah.Mvc" version="2.1.2" targetFramework="net452" />
-  <package id="Hangfire" version="1.6.22" targetFramework="net472" />
+  <package id="Hangfire" version="1.7.5" targetFramework="net472" />
   <package id="Hangfire.Console" version="1.4.2" targetFramework="net461" />
-  <package id="Hangfire.Core" version="1.6.22" targetFramework="net472" />
-  <package id="Hangfire.SqlServer" version="1.6.22" targetFramework="net472" />
-  <package id="jQuery" version="3.3.1" targetFramework="net461" />
+  <package id="Hangfire.Core" version="1.7.5" targetFramework="net472" />
+  <package id="Hangfire.SqlServer" version="1.7.5" targetFramework="net472" />
+  <package id="jQuery" version="3.4.1" targetFramework="net472" />
   <package id="Lucene.Net" version="3.0.3" targetFramework="net452" />
   <package id="Lucene.Net.Contrib" version="3.0.3" targetFramework="net452" />
   <package id="Microsoft.AspNet.Cors" version="5.2.7" targetFramework="net472" />
@@ -25,33 +25,34 @@
   <package id="Microsoft.AspNet.WebPages" version="3.2.7" targetFramework="net472" />
   <package id="Microsoft.CodeDom.Providers.DotNetCompilerPlatform" version="2.0.1" targetFramework="net461" />
   <package id="Microsoft.CSharp" version="4.5.0" targetFramework="net472" />
-  <package id="Microsoft.EntityFrameworkCore" version="2.2.2" targetFramework="net461" />
-  <package id="Microsoft.EntityFrameworkCore.Abstractions" version="2.2.2" targetFramework="net461" />
-  <package id="Microsoft.EntityFrameworkCore.Analyzers" version="2.2.2" targetFramework="net461" />
-  <package id="Microsoft.EntityFrameworkCore.Design" version="2.2.2" targetFramework="net472" />
-  <package id="Microsoft.EntityFrameworkCore.Relational" version="2.2.2" targetFramework="net472" />
+  <package id="Microsoft.EntityFrameworkCore" version="2.2.6" targetFramework="net472" />
+  <package id="Microsoft.EntityFrameworkCore.Abstractions" version="2.2.6" targetFramework="net472" />
+  <package id="Microsoft.EntityFrameworkCore.Analyzers" version="2.2.6" targetFramework="net472" />
+  <package id="Microsoft.EntityFrameworkCore.Design" version="2.2.6" targetFramework="net472" />
+  <package id="Microsoft.EntityFrameworkCore.Relational" version="2.2.6" targetFramework="net472" />
   <package id="Microsoft.Extensions.Caching.Abstractions" version="2.2.0" targetFramework="net461" />
   <package id="Microsoft.Extensions.Caching.Memory" version="2.2.0" targetFramework="net461" />
   <package id="Microsoft.Extensions.Configuration" version="2.2.0" targetFramework="net461" />
   <package id="Microsoft.Extensions.Configuration.Abstractions" version="2.2.0" targetFramework="net461" />
-  <package id="Microsoft.Extensions.Configuration.Binder" version="2.2.0" targetFramework="net461" />
+  <package id="Microsoft.Extensions.Configuration.Binder" version="2.2.4" targetFramework="net472" />
   <package id="Microsoft.Extensions.DependencyInjection" version="2.2.0" targetFramework="net461" />
   <package id="Microsoft.Extensions.DependencyInjection.Abstractions" version="2.2.0" targetFramework="net461" />
   <package id="Microsoft.Extensions.Logging" version="2.2.0" targetFramework="net461" />
   <package id="Microsoft.Extensions.Logging.Abstractions" version="2.2.0" targetFramework="net461" />
   <package id="Microsoft.Extensions.Options" version="2.2.0" targetFramework="net461" />
   <package id="Microsoft.Extensions.Primitives" version="2.2.0" targetFramework="net461" />
-  <package id="Microsoft.IdentityModel.JsonWebTokens" version="5.4.0" targetFramework="net472" />
-  <package id="Microsoft.IdentityModel.Logging" version="5.4.0" targetFramework="net472" />
-  <package id="Microsoft.IdentityModel.Tokens" version="5.4.0" targetFramework="net472" />
-  <package id="Microsoft.Net.Compilers" version="2.10.0" targetFramework="net472" developmentDependency="true" />
+  <package id="Microsoft.IdentityModel.JsonWebTokens" version="5.5.0" targetFramework="net472" />
+  <package id="Microsoft.IdentityModel.Logging" version="5.5.0" targetFramework="net472" />
+  <package id="Microsoft.IdentityModel.Tokens" version="5.5.0" targetFramework="net472" />
+  <package id="Microsoft.Net.Compilers" version="3.1.1" targetFramework="net472" developmentDependency="true" />
   <package id="Microsoft.Owin" version="4.0.1" targetFramework="net472" />
   <package id="Microsoft.Owin.Host.SystemWeb" version="4.0.1" targetFramework="net472" />
   <package id="Microsoft.Web.Infrastructure" version="1.0.0.0" targetFramework="net461" />
   <package id="Microsoft.Win32.Primitives" version="4.3.0" targetFramework="net46" />
   <package id="Modernizr" version="2.8.3" targetFramework="net461" />
-  <package id="Newtonsoft.Json" version="12.0.1" targetFramework="net472" />
-  <package id="NJsonSchema" version="9.13.18" targetFramework="net472" />
+  <package id="Namotion.Reflection" version="1.0.5" targetFramework="net472" />
+  <package id="Newtonsoft.Json" version="12.0.2" targetFramework="net472" />
+  <package id="NJsonSchema" version="10.0.21" targetFramework="net472" />
   <package id="Owin" version="1.0" targetFramework="net461" />
   <package id="popper.js" version="1.14.3" targetFramework="net472" />
   <package id="Remotion.Linq" version="2.2.0" targetFramework="net461" />
@@ -66,14 +67,14 @@
   <package id="System.Console" version="4.3.1" targetFramework="net472" />
   <package id="System.Diagnostics.DiagnosticSource" version="4.5.1" targetFramework="net472" />
   <package id="System.Globalization.Calendars" version="4.3.0" targetFramework="net46" />
-  <package id="System.IdentityModel.Tokens.Jwt" version="5.4.0" targetFramework="net472" />
+  <package id="System.IdentityModel.Tokens.Jwt" version="5.5.0" targetFramework="net472" />
   <package id="System.Interactive.Async" version="3.2.0" targetFramework="net461" />
   <package id="System.IO" version="4.3.0" targetFramework="net472" />
   <package id="System.IO.Compression" version="4.3.0" targetFramework="net46" />
   <package id="System.IO.Compression.ZipFile" version="4.3.0" targetFramework="net46" />
   <package id="System.IO.FileSystem" version="4.3.0" targetFramework="net46" />
   <package id="System.IO.FileSystem.Primitives" version="4.3.0" targetFramework="net46" />
-  <package id="System.Memory" version="4.5.2" targetFramework="net472" />
+  <package id="System.Memory" version="4.5.3" targetFramework="net472" />
   <package id="System.Net.Http" version="4.3.4" targetFramework="net472" />
   <package id="System.Net.Sockets" version="4.3.0" targetFramework="net46" />
   <package id="System.Numerics.Vectors" version="4.5.0" targetFramework="net472" />
