--- CSETWebApi/CSETWeb_Api/CSETWeb_ApiTests/packages.config
@@ -1,11 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="EntityFramework" version="6.2.0" targetFramework="net452" />
-  <package id="Microsoft.AspNet.WebApi" version="5.2.6" targetFramework="net46" />
-  <package id="Microsoft.AspNet.WebApi.Client" version="5.2.6" targetFramework="net46" />
-  <package id="Microsoft.AspNet.WebApi.Core" version="5.2.6" targetFramework="net46" />
-  <package id="Microsoft.AspNet.WebApi.WebHost" version="5.2.6" targetFramework="net46" />
-  <package id="MSTest.TestAdapter" version="1.3.2" targetFramework="net46" />
-  <package id="MSTest.TestFramework" version="1.3.2" targetFramework="net46" />
-  <package id="Newtonsoft.Json" version="11.0.2" targetFramework="net46" />
-</packages>
